Allowing inverting -/+ symbols in transforms UI
Users had to move the cursor to te beginning of a number in a transform and type `-` to make the number negative. That was a bit annoying so I changed the logic to invert the current sign of a number by typing `-` anywhere in the field.

I have also removed a hack for specify min field size with decimals... This is handled in CSS better now.