Adding ability to invert a 4x4 matrix
In animation, sometimes we need the inverse of a matrix to figure out how much rotation/translation/scale has been done on an object. This is needed when the transforms occur in a different space so we need to extraprolate only what changed. This is how bone animation works, especially with pose matrices.

The changes here add the ability to calculate the determinant and the adjoint of a matrix to invert it. The new methods are `invert`, `adjoint`, and `determinant`.