Refactored matrix for better memory management
Matrices we creating new arrays and coercing values every single time a new matrix was created, which occurs on every single transform operation. This is isn't optimal.  So I refactored the code to have:

1. a getter for Matrix.data. This is where we now do the coercion for floats instead of every operation
2. Expensive operations like multiplication and rotation that were creating a new array for everything are now able to take a destination array so that for each transform that calls multiple things the same final array can be shared.

The total net effect might be neglible but because we are not creating as many arrays we are more gentle on gargabe collection.