Refactored matrix inverting code to be more efficient
For simplicity purposes, matrix inverting code was calling adjoint and determinant to calculate the invert of the matrix. But that was not as efficient as it could have been because both of those function were doing a lot of overlapping work.  So I folded the common logic and Im just doing all the work in the invert function.  There is duplicated code, but that's fine.