Concat all the rotation matrices into a single matrix
Previously I was rotating one axis at a time and multiplying results together. This was more expensive than it needed to be, so I added a single rotation matrix with all the things done in a single step.

The single rotation matrix was generated with a new helper function that can multiply multiple matrices together where the matrices have variable names.  The out is a matrix with all the variable names that you can use in your code.