Changed transform matrices to have translation the last column
So I used to have all the transform matrices with the translation vector in the last row.  But this required everything to always be transposed which caused a lot mental friction and overhead testing new functionality because math literature was reversed and was difficult to map things like rotation matrices. It was very limiting with tools that exist online for testing matrices as well. It was also very unintuitive because the order in which matrix multiplication had to occur was reversed! The whole mental model was revered. The last straw was that I just couldnt easily map to which notation (Classic euler or Tait–Bryan) I was using. And then implementing quaternions... Everything was backwards and that is hard to debug. The one drawback of this change is that now matrices need to be transposed when passed in to vertex shaders where we multiple the view matrix with the vertex vector like `gl_Position = view * position`.  If your shader is position first, then you dont need to do the transpose. So there is a small amount of friction but it is a good compromise.

This change will also make it simpler to implement quaternions without having to think about subtlties with having to work with matrices transposed.

So I finally decided to be more friendly with literature and tools over webgl.