Matrix4 now uses Math.sin and Math.cos directly
I have wrappers around Math.sin and Math.con type of functions that used to have optimizations where values were in a look up table. They also automatically converted degrees to radians.  But that's not as useful anymore for the general case so Im changing the main consumer of that (Matrix4) to use Math.<X> functions directly.