Adding fbx animation test to verify tricky rotation
Im adding a test based on cubearmature_simple.fbx designed to verify that rotation propagates from bones to skins correctly. The more tricky bit is correctly rotating the Right cube on the Y axis 90deg after the Right cube bone is rotated on the X axis 90deg. This verifies issues like gimbal lock and that rotations propagate correctly in a scene graph.