Changed FOV in the perspective matrix for the editors scene
Previously the FOV for the perspective projection matrix was 90 degrees, which causes object towards the edge of the center of whats being rendered to look stretched (distorted), which didnt look great when viewing the object up close because most of it would look stretched.  The changes here change the FOV to 35 degrees, which reduces the stretching feels significantly.  This caused me to update all the different FBX file to not be so small.  All exports were about 0.3 in size of the original size vs 0.01.  I also had to update one of the tests for armature animation.