Adding animation pause/resume functionality
Changes in this diff add the ability to start, pause, and resume animation. You can get at this from the animation node in a model or the animation panel of the model itself.

One thing Im including in these changes is the addition of iconoir-react, which is available at https://www.flaticon.com/packs/library.  Im adding the icon for play/pause. I will be using more of the icons for other parts of the UI.

Next, Im adding the ability to skip frames (forward and backward).