Refactored keyframe animation to use classes for better access to animation state
Im working on adding the ability skip frames in animation forward and back as well as fast forwarding. And to do that I need better access to animation state so I have refactored animation keyframe controllers to be classes with helpers to access animation state like frame indexes, frames, and times internals.