Adding ability to skip frames in FBX animations
Im working on support for FBX keyframe animation and I am the ability to skip frames. On the cosmetic side of things, the changes add rounded buttons to skip forward and backwards. I have also updated the styling of the play/pause buttons to be rounded buttons.

On the logic side of things. Skipping frames is currently based on the assumtion that every second has 24 frames.  So every time the user presses the button we basically advane the animation timer by 1000/24 where 1000 is 1 second and 24 is the number of frames per second.  I have also moves the animation playback functionality to stack nodes and changes the logic so animatable nodes go thru the active stack to read the playback time to evaluate animation frames.