Fixed issue with FBX Cluster Deformers being direct children of geometry
Previously the assumption was that Cluster Deformers in FBX files could only be children of Skin Deformers, but that's not entirely true as I have found files that have Cluster Deformers as direct children of Geometry nodes. So I have changed the logic here to ensure that we traverse FBX connection upwards finding the correct ancestor node (Geometry node) that has a `polygonIndexMap` defined.