Tweaked FBX node names once again!
Node names _always_ incremented a counter and always had a suffix of `_n<index>`. But for the most part the more relevant nodes are unique, so the logic for generating names are that we only start adding suffixes when we find duplicates, in which case we will generate names in the for `_n<index>` where `index` starts at 1.

So for repeated names it would like:

Geometry
Geometry_n1
Geometry_n2
...
Geometry_n<index>