Adding TRS function to Matrix4
Usually when creating a new matrix we provide the translation, rotation, and scale information. Thus fat the simplest way to create a matrix with that information is by calling `Matrix4.translation(...).rotation(...).scale(...)`. But that overhead that was not needed from multipliying all of those operations to correct aggregate them. To optimize that I have added a function `trs` that takes the same information as separate arrays and it will avoid creating redundant arrays and multiplication operations.

I have also reorganized the matrix tests a lil bit.