Refactored logic for getting the animation info in an animatable scene node
I moved a bunch of logic that was in the preRender function that was for calculating the animation frame for an animatable object. That's to hopefully declutter the preRender function and also make it so that the animation can be calculated by any logic as long as there is an animatable node.