Refactore where and how gl rendering logic happens to fix memory leaks and logic
Since adding FBX support, different scene nodes like textures have needed to add their own attributes and uniforms to gl shaders via the ShaderProgram objects. But the nodes resposible for calling the render functions with vertex objects and shader programs wasn't correct cleaning up these attributes and uniforms that were added in each render cycle (about 60 times a second) which was causing huge memory leaks.

So I have fixed the memory by cloning the shader program in every preRender call and ensuring we copy the shader program used when nodes were created so that we can reset all the attributes and uniforms that are added in each render call.

To make the code much cleaner, I have also moved all the render functionality to ShaderProgram so that it is easier to see what's involved in rendering each frame. And this allowed me to use the `gl` context from the shader programs instead of the `gl` context in the render context. In addition to that, I am also passing the `gl` context from the shader program to the update functions for gl uniforms.

I will be adding attribute and uniform caching tho, which further clean up the need to clone the shader program every time we render a frame which will help us avoid creating lots of array in every frame.