Adding shader uniform caching
Previously I had made changes to ensure that shader programs have a clean slate each time a frame is rendered to fix memory leaks. That was also setting things up for the changes in this commit where we cache shader uniforms to avoid recreating them in every frame.  I have also split out of the Uniform class the update function to keep separate state updates and uniforms data to avoid holding onto memory that is not needed by uniforms and avoid the need to call `getUniformLocation` in every frame.

The new semantics keep a map by uniform name where we store uniform indexes so that we don't have to call `getUniformLocation` in every frame. And we also have a map by uniform name for the updater functions the shader program calls for providing uniform state to the shader program.