Moved logic to register bones to the armature node itself
We used to register bones in a skeleton animation in bones and armature nodes as bones were added in the scene. This was confusing because if a bone was not a desendent of armature then the registration process in the bone would not happen.  I had to add comments to clarify when a bone wouldnt have an armature as a parent which helped, but it was ultimately confusing.  So I have consolidated the logic to register bones in the armature itself which makes it super clear that the registration process 100% depends on there being an armature as enforced by the API itself. Bones are registered lazily the first time that `bonesByID` in the armature is accessed.