Renamed reindexPolygonVertex to polygonVertexIndexToDirect in the FBX file processor
FBX has a lot of indexes used for different situations and this can be very complicated to keep straight in your head.  So I am renaming some functions to better communicate in terms of FBX file format what we are converting from and to. In this case I have renamed `reindexPolygonVertex` to `polygonVertexIndexToDirect` because the former doesn't really clearly communicate what is being reindexed and to.