Adding support for vec4 attributes parsing in the shader factory
For skin bone animation, I am going to need to specify vec4 attributes in the vertex shaders for bones weights and bone matrix indexes. The changes here adjust all the vertex shaders to have the correct vector size, add support for vec4 in the shader factory, and do better checks when attribute sizes as defined in the vertex shaders and vertex buffer don't match.