Disabled auto animation resetting to get animation tests working again
When I added the playback abstraction I changed the logic so that animation would reset the playback for an animation when transitioning between stacks (tracks). This was causing animation tests to fail because animation timer would no longer advance.

So Im disabling it for now to make sure we have passing tests and we will figure out a different way to reset playback when switching animation.

I have also moved all the files that I use for testing to __testdata__ folder to make it less likely to modify those files which can be sensitive to exporting with the wrong parameters.