Adding support for rigged animation with weighted vertices
I have finally implemented support for rigged animation with skin deformers. The information for this is encapsulated in clusters, which are scene nodes that contain the indexes for the vertices they affect and the weight, which is how much they affect a vertex in relation to other clusters.  In a way, clusters are like the vertex meta data related to a bone in rigged animation. Each cluster has the ID of the bone they correspond to so when the armature goes thru the preRender phase, clusters can get the world transformation for the bones itself so that the vertices a cluster affects can be rendered in world space.

For this whole thing to work correctly, I gather all the weights for each bone, assign each cluster some unique index used by the vertex to know which animation matrix corresponds to each groups of vertex indexes.  The animation matrices (the bone matrices) are provided to the vertex shader as a uniform. And the weights and bone matrix indexes are both provided as buffer object to the vertx shader.

Skin animation is based on having multiple bones affect any particular vertex by some amount. We are limiting the number of bones that can affect a vertex to 4, which is quite common. When a vertex s influenced by more than 4 bones (clusters), I will pick the 4 with the highest weights (with the most influence) on any particular vertex.

In the vertex shader, I have picked something like 80 bone matrices that can be provided. It's a relatively arbitrary number i have chosen based on the number of total bone matrices I saw in a few different models. this might be sufficient - not sure. we are just going to test it for a lil and see if we need more or less.