Reworked attribute binding from shader-factory to shader/program
The editor had all the logic in the shader-factory file for dealing with creating shader programs and auto binding attributes. I have moved all that logic to the shader program file itself to enhace the built in funcitonality for the shader program factory.

I have also updated the documentation for new methods and updated methods around shader program creation.  I have left the caching and minimum functionality for async loading shader sources in the editor since that seems to be more specific to how the editor itself works.  But it is a very thing layer that should be easy to replicate for different contexts.