Reporting of compilation errors for shaders
We never really had any indication about failures to compile a shader with `gl.compileShader`, which made it more difficult to know what shaders had issues befor using them. So you would instead end up with obscure errors when binding uniforms and attributes. The changes here introduce some baseline parsing of errors that come up when compiling a shader, we will report them along with the line that caused the compilation error to begin with.

In a follow up commit I will be adding better handling for attributes in a shader that are not used. Right now we just throw up a huge red screen without a ton of information them, and usually not using an attribute is not a reason to stop the entire shader from doing its work.

I have also fixed linting error from missing // eslint-disable-next-line no-console