Report warnings for unused shader attributes
Previously when we vertex shaders had attributes that weren't used, the index would be -1.  In those cases we would just throw an exception which would cause the entire scene to be covered with a red screen complaining about it.

I am changing the logic to be more gentle and instead of throwing an exception, I am just logging a warning, make sure that attribute isn't added to the shader program, and will let the program continue on to do its work.