Adding ability to select new FBX files in an existing scene node
From the scene tree, you can select a static mesh node with an already loaded FBX file and load a different file. To do this correctly, I have added the ability to clear a node's children. And I have also removed the "hack" where we store the animation node as a property of the static mesh node that contains the FBX model.