Fixed issue parsing gl errors without line numbers
When we compile a shader I check for failures, which usually have a line number because they are for syntax errors.  But the other class of errors like _system_ errors do not include line number information and this was causing failures.  So these changes change the regex to skip the line number section if no line numbers exist in the error messages.