Adding ability to get details about webgl and show them when errors occur
I have added a new function called getDebugData(gl) which reads several bits of useful debug details from a webgl context. I am getting `WEBGL_debug_renderer_info` for getting the `vendor` and the `renderer`. Im also getting limits and context attributes.  All this information is added in the console logs we show when errors are reported in the UI.

If you want to get at all sorts of extra details, go to the URL in your browser
https://webglreport.com/