Bone matrices texture for skinned animation
In skinned animation we were preallocating 64 bone matrices in the vertex shader, which was too much for safari breaking rendering. Safari was happier with 61 bone matrices but that was too few for some models, which caused rendering to break because  only partial trasnforms were applied.  What made this bad is that Chrome worked just fine with 64 (or even more) preallocated bone matrices, so it was either support Safari and not support more complex animated models or limit the total number of bones, which would have caused other issues because we were just running into max number of uniforms in a shader.

To fix those constraints I have implemented the approach to store matrices in a texture which we read data from in the vertex fragment via texelFetch.  By storing he matrix information in these textures we entirely avoid all the issues with max uniform count.

There is one caveat that is worth highlighting. And that's the textures are automatically flipped on its Y axis because images are upside down.  To address this I just reverse the list of bone matrices before creating the texture.