Fixed issue with animation weighted matrices and normal vectors
When animating, we were still applying the world matrix without animation to the normal vectors causing lighting to not work correctly; it looked like light sources were not lighting the correct part of the models when animation was applied.  To fix it Im just using the weighted animation matrix in place of the static world animation.