Adding FBX animated character as default scene model
Now that skinned (rigged) animation is stable, I have added an animated character to the default scene to demostrate how that works in the editor.  The characted is from https://www.mixamo.com/ and I just downloaded:

1. cat walk
2. twerk
3. salsa
4. jazz

And combined those animations into a single character so that we can have multiple animations to choose from. The files directly from Mixamo are causing issue; I can't tell if its because the FBX format version is 7700 or the FBX file is using unsupported features like pre rotation matrices.  I will look at this separately because ideally Mixamo characters would work out of the box.

To test, just load localhost:3000/editor/index.html and select the animation node on the tree. Then press the play button to run an animation or select a new one from the drop down.