Adding skinned mesh scene node
After having implemented skinned animation for FBX files, I am adding a new core scene node called `SkinnedMesh`. This node has directly built in logic for storing animation scene nodes for easy access. This is now wired up in the editors tree view and detail panels to show `Animation` right in the mesh node. This makes it easier to work with the tree because we don't have to expand the mesh node and find the animation in the tree itself in order to run an animation; you access the animatio panel right from the mesh details panel.