Fixed a few css issues to add better vertical scrolling
The three and the details panel were getting all smooshed and it was hard to scroll in smaller size screens that arent so tall. This caused UX problems in general.  So Im just tweaking a few things to improve scrolling when the tree is expanded and detail panels are shown.