Returning null when node details panel doesnt have a valid scene node
We had an issue with node details panel getting rendered taking up space in the scene tree but there was no node to render, so the scene tree was rendering only half size.  To fix that Im just not rendering any details panel at all when the node to be rendered is not known.