Moving the Material fbx scene node as a core scene node
Moving the Material node from the fbx custom node to core so that we can allow other types of renderable nodes to leverage it. In a follow up commit I will be wiring up a panel to it so that folks can change the color of a mesh thru that.