Adding support for ByVertice in geometry layer for FBX files
While testing animated models from Mixamo I found that they use ByVertice quite often for Normal vector geometry layers, so I added support for it. The work around was to re-export models with blender but we don't need that anymore.

Context:
Vertices in an FBX model are indexed for building polygons which is why vertex indexes are called `polygonVertexIndex`. ByVertice means that geometry layers have indexes that map directly to vertex indexes sequentially (vertex index 0 maps to normal index 0) and not polygon indexes. Polygon indexes are groups of vertex indexes that we iterate to build triangles (and quads). So we remap ByVertice to indexes we can use for rendering using the polygon vertex indexes.

https://banexdevblog.wordpress.com/2014/06/23/a-quick-tutorial-about-the-fbx-ascii-format/