Changed exception to a log message when creating key frame animator with just 1 frame
Testing mixamo models I found that some created animation with just 1 frame. Clearly there is nothing to animate here, but we still want to handle it gracefully.  So I have changed the exception we used to throw to a console log so that we know something is a lil off but not necessarily broken.