Adding support for node.id as unique identifiers
For the longest time all the nodes heavily relied on node names to be unique. This served as a way to easily create scenes directly from scene configurations.  I had added small guard rails for avoiding name duplication in the model loaders, but that was not an reliable enough guarantee that names would be unique.  Also if somehow two nodes ended up with the same name, selecting one of them in the scene tree in the editor would cause both to be highlighted. Using names also made more difficult to support editor the name of nodes in a scene editor!

So too many short comings so I am adding support for unique IDs in scene nodes.  You can entirely omit `id` from the constructor and a unique one will be generated for you, or you can pass an `id` value and we will ensure it is unique before using it.

Next changes are for propagating the use of ID instead of node names.