Adding support for reading scene nodes by ID in a scene
We have been relying on the name of a scene node to find nodes in a scene. But that has many drawbacks like the name of the node needed to be unique or things would break, renaming a node via the editor could be problematic, and semantics around ID are just overall more stable because that can't really be changed.

To do this, I had to change the flow of how things are loaded in the editor and had to adjust the buildDefaultState to inject an ID into the configuration before the scene nodes were created.  This process was also moved out of the `createScene` function and moved up in the stack when the editor is being initialized.  That was done so that the update scene config can be used for creating the scene graph and for the resource loader to have access to the IDs of the nodes that were created, which in the passed just relied on the name specified in the scene config.

That allowed us to also have more clear logic in the `buildSceneNode` used in the resource loader because the resource loader can now verify and ensure that the scene node exists before it is provided to the `buildSceneNode` function.  A side effect of this is that `buildSceneNode` can no longer return a custom scene node that directly corresponds to a node specified in the scene config `type`field, which is OK because that wasn't really being used and semantics weren't too clearly defined. So `buildSceneNode` now always gets a scene node that was created during the `createScene` call.

The next thing to do to close the loop with this is to update the scene manager to use the ID of the node as a key instead of the name. This will be done in a follow up commit.