Adding prettier to start automatically formatting scenic code
Initial configuration for prettier.  This is adding configuration files with default settings, ignore everything but editor/*.js to start with, and prettified README.md to try it out.

To run prettier right now we just run `npx prettier --check .` to check for what needs updating.  And `npx prettier --write .` to fix issues.