Rendering tabs in for skinned animation nodes without animations crashes
I saw a crash that happened when loading the `JediStarFighter.fbx` model in the `rigged animation` node. And the reason is that there is no animation in that model so the toolbar in for the skinned animation panel was sending the toolbar `null` in of the animation tab value.  The changes in this commit simply filter out falsy values.