Fixed issue with Obj files not loading when added thru the editor
The issue happened when we were loading an Obj into a skinned scene bone but the logic in the Obj loader was missing the check for skinned models.  I have also added logic to make sure scene nodes get all their state and data cleared before loading a new error.