Renamed childrenByType to childrenByTypeName
Nodes have a `type` field which has a string that is used for grouping nodes by type. The name `type` however, can be confusing because that's also used for `instanceof` checks.  To make the difference more clear I renamed the uses of `type` fields to type name. That should make it more clear that we are at least dealing with string values.