refactored logic so that mesh render can be called by any node (#12)
previously the `Mesh` class defined a member render function that inheritors of the mesh class called to render mesh data. But I started to run into some limitations beucase a scene node that needed to be animated also needed to inherit from Animatable. So i changed the Mesh render function a static function so that it can be called without inheriting.  This provides a much cleaner separation of concerns too because previously several classes that were not mesh inherited from mesh for convenience to access the renderable interface.