Adding better organization and composability of Mesh class
I have been working on fine tuning the API I want for composing the Mesh api for rendering in a way that makes it very clear which FBX animatable classes need mesh functionality and which ones don't.  So I implemented the AnimatableInterface that allows me to extend animatable functionality with different super classes.  And from the mesh standpoint, I am enabling inheriting classes to override the way in which materials and lights state are calculated for rendering the mesh.  With this combination, the FBX mesh can create an animatable mesh that can specify a way to configure materials state which uses the values from the relative root instead of the mesh itself.  The reason we use the relative root is that is where the scene configurations (scene-config.js) exist, and those are the values that the scene tree affect in the UI.