Adding loadScene function to load a new scene
Im working on adding support for loading a new scene from the editor. And to do that I have added a new `loadScene` function that takes a scene configuration object. You can load up a JSON file and feed that loadScene the result to replace the current scene with a new one.

We will have follow up change in the UI.