Doing a quick POC for importing CSS files in JSX components
In #16 i have added a Panel component with support for css style objects.  but the question of whether or not we should use imports was raised, which I think is a very viable path forward as well.

This use built in pakit support for importing CSS which currently doesn't support scoping styles. But Im going to work on adding that as an optin feature.  In the meantime what we have is good enough to try out css imports.