Migrated all selectedView state to hooks (#20)
* Adding Panel, PanelHeader, and PanelBody components

Im refactoring the UI to use a more componentized. Low hanging fruit is switching things to use a panel. This commit only does one component but a follow up commit will do the rest of panel migration.

* Migrated all selectedView state to hooks

The scene editor has a lot of panels with tabbed panel semantics. We kept selectedView in the component state but that had a lot of boilerplate code for setting things up.  Moving that to just using useState hook cleans things up significantly.