Adding FBX scpecific Animation node
So we have been just using the Animation node from scenic core which has no logic, and instead other FBX scene nodes were responsible for managing the current animation stack which made the code harder to think about. It was also very difficult to support resetting an animation stack starting time when a new animation stack was selected.

In order to address those issues/sharp edges, Im adding a new FBX Animation node. It will manage the current stack so that animatable objects dont have to worry about that. I have also been able to reset animation stacks when a new is selected. And this will also enable the ability to get information about the current more easily so that the UI can know the animation stack length for scrubbing support.