Fixing tests broken when adding FBX Animation class
In a previous I broke the fbx loader test because I changed where animation stacks are stored now that we have an FBX Animation class. I also tweaked the animation test becaue we werent creating a skinned-mesh object which sets the correct animation as a property of the mesh, so the logic for animation needed to find the animation node which wasn't as efficient.