fixed range function inconsistent handling of decimal points
range is a number generator function that takes a start and an end for a range. The length of the range is calculate based on the start/end pair, but that sometimes leads to the wrong length calculations when dealing with ranges of decimal numbers.

For example `range(0, 0.9, 0.1)` generates a range from 0 to 0.9. But `range(1, 1.9, 0.1)` generates a range from 1 to 1.8. The missing 1.9 in the range is because the calculations for length generates the wrong number. To handle the length generation more consistently, Im using `Math.round` which does yield the correct length in those edge cases and the range from 1 to 1.9 now includes the 1.9 as expected.