Changed the animation properties panel to convert speed strings to numbers
The animation properties panel uses an input for changing the speed of an animation track, and that html input generates strings instead of numbers. It works just fine passing those values around as strings because number multiplication coerces them to number. Instead of doing that magically, im converting the string to number as they speed numbers change in the html input.