Fixed issue calculation animation delta with varying time ranges
Some animations specify variables time ranges for frames, which is useful when you want different frames to last more or less times. And that complicates the calculations for delta values in animation interpolation.  The changes here basiacally fixed the logic because the values just werent correct, but they were good enough that it was hard to notice anything wrong when playing animation.  At worst there were very subtle jiggers in the animation objects.

I have also split out the logic for wrapping time so that we can test that separately.