Normalize time intervals in FBX file animations
FBX animation defines a second as being `46186158000`.  So every time we are calculating animation data we have to multiply regular javascript ms times that number so that they match up with the time intervals in an animation track.  There was a bug that made normalizing animation data to remove `46186158000` which I fixed here https://github.com/MiguelCastillo/scenic/commit/4ea14bd51fdad422c76d00636ad7ab862a2331c5.  That used to cause weird jitters due to bad delta values causing noise when LERPing between frames.  But now that fixed so the changes here are to pre-process time intervals in FBX animations so that we no longer have to `46186158000` when calculating animation data at runtime.